pyproject.toml
[project]
name = "bodytest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.32.4",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "asyncio>=3.4.3",
]
